requirements.txt
eel
eel[jinja2]
pyyaml
pyautogui